% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamsInit.R
\name{cmm_init}
\alias{cmm_init}
\title{Initialize an LRMoE model based on Clustered Method of Moments}
\usage{
cmm_init(Y, X, n_comp, type, exact_Y = FALSE)
}
\arguments{
\item{Y}{A N by d (\code{exact_Y=T}) or N by 4d (\code{exact_Y=F}) matrix of numerics,
where N is sample size and d is the dimension of each obsevation.
If the size is N by 4d, Each block of four columns should be organized as \code{(tl, yl, yu, tu)}, representing the
truncation lower bound, censoring lower bound, censoring upper bound and truncation upper bound.}

\item{X}{X A N*P matrix of numerics, where P is the number of covariates.
The first column of \code{X} should be 1, which is the intercept.}

\item{n_comp}{Numeric, representing how many latent classes/groups to use.}

\item{type}{A vector of strings of either "continuous" or "exact", representing
whether each dimension of \code{Y} is continuous or discrete.}

\item{exact_Y}{TRUE/FALSE: whether \code{Y} is observed exactly, or with censoring and/or truncation.}
}
\value{
A list where \code{zero_y}, \code{mean_y_pos}, \code{var_y_pos},
\code{skewness_y_pos} and \code{kurtosis_y_pos} represent the summary statistics of \code{Y}
by dimension and by component.
\code{alpha_init} and \code{experts_init} represents the parameter initializations.
\code{ll_init} contains the loglikelihood of each experts fitted to \code{Y} by
dimension and by component.
\code{ll_best} suggests an initialization of expert functions based on the best loglikelihood.
}
\description{
Initialize an LRMoE model based on Clustered Method of Moments
}
