% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMMInitParams.R
\name{CMMFrequency}
\alias{CMMFrequency}
\title{Initializes parameter for frequency distributions using CMM.}
\usage{
CMMFrequency(Y, X, n.comp)
}
\arguments{
\item{Y}{A vector of response variables.}

\item{X}{A matrix of covariates (normalized for running \code{\link[stats]{kmeans}}).}

\item{n.comp}{Number of components}
}
\value{
A list of parameter initialization.
}
\description{
Initializes parameter for frequency distributions using CMM.
}
