% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMMInitParams.R
\name{CMMInit}
\alias{CMMInit}
\title{Initializes parameter for severity/frequency distributions using CMM.}
\usage{
CMMInit(Y, X, n.comp, type = NULL)
}
\arguments{
\item{Y}{A matrix of response variables.}

\item{X}{A matrix of covariates (normalized for running \code{\link[stats]{kmeans}}).}

\item{n.comp}{Number of components}

\item{type}{A vector of strings, either 'S' for severity or 'F' for frequency, corresponding to each dimention of \code{Y}.}
}
\value{
A list of parameter initialization.
}
\description{
Initializes parameter for severity/frequency distributions using CMM.
}
