% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMMInitParams.R
\name{CMMSeverity}
\alias{CMMSeverity}
\title{Initializes parameter for severity distributions using CMM.}
\usage{
CMMSeverity(Y, X, n.comp)
}
\arguments{
\item{Y}{A vector of response variables.}

\item{X}{A matrix of covariates (normalized for running \code{\link[stats]{kmeans}}).}

\item{n.comp}{Number of components}
}
\value{
A list of parameter initialization.
}
\description{
Initializes parameter for severity distributions using CMM.
}
