% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountParams.R
\name{CountZero}
\alias{CountZero}
\title{Count the number of zero probability masses.}
\usage{
CountZero(comp.dist)
}
\arguments{
\item{comp.dist}{A d*g matrix of strings describing the component distributions.}
}
\value{
The number of zero probability masses.
}
\description{
Count the number of zero probability masses.
}
\keyword{internal}
