% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaCountUtils.R
\name{dgammacount}
\alias{dgammacount}
\title{Modified \code{\link[rmutil]{GammaCount}} pmf for better numerical accuracy.}
\usage{
dgammacount(y, m, s, log = FALSE)
}
\arguments{
\item{y}{Vector of gamma count values.}

\item{m, s}{Paramaters of Gamma count distribution.}

\item{log}{TRUE/FALSE: whether log density should be returned.}
}
\description{
Modified \code{\link[rmutil]{GammaCount}} pmf for better numerical accuracy.
}
\seealso{
\code{\link[rmutil]{GammaCount}}.
}
