% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimCompExpertLL.R
\name{DimCompExpertLL}
\alias{DimCompExpertLL}
\title{Computes loglikelihood with zero inflation, by dimension and by component.}
\usage{
DimCompExpertLL(Y, comp.dist, zero.init, params.init)
}
\arguments{
\item{Y}{An \code{N*4d} matrix of numerics, where N is the sample size and d is the dimension of observation.
For each dimension, the four columns shoud be organized as \code{(tl, yl, yu, tu)}.}

\item{comp.dist}{A \code{d*g} matrix of strings, where d is the dimension of observation and g is the number of components.}

\item{zero.init}{A \code{d*g} matrix of numerics between 0 and 1, representing zero inflation by dimension and by component.}

\item{params.init}{A list of length d. Each element in this list is again a sublist of length g.
Each sublist contains one vector of numerics, representing the parameters of experts by dimension and by component.}
}
\value{
\code{dim.comp.dist}: Same as input \code{comp.dist}, for passing to other functions calling this function.

\code{zero.inflation}: A d*g matrix of TRUE/FALSE, where TRUE indicates the expert is zero-inflated.

\code{zero.prob}: Same as input \code{zero.init}.

\code{pos.params}: Same as input \code{params.init}.

\code{pos.expert.ll, pos.expert.tn, pos.expert.tn.bar, expert.ll, expert.tn, expert.tn.bar} See \code{\link{PosExpertLL}}.
        Here, \code{pos.} indicates the values are for positive parts only, while other values are for zero-inflated versions.
}
\description{
Computes loglikelihood with zero inflation, by dimension and by component.
}
\keyword{internal}
