% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLimEx.R
\name{DimCompPosYLimEx}
\alias{DimCompPosYLimEx}
\title{A function to calculate the limited mean of a matrix of given distributions (positive part only), by dimension and by component.}
\usage{
DimCompPosYLimEx(comp.dist, params.list, limit)
}
\arguments{
\item{comp.dist}{A d*g matrix of strings, describing component distributions by dimension and by component.}

\item{limit}{A vector of limit values by dimension.}

\item{paramas.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A matrix of limited mean values by dimension and by component.
}
\description{
A function to calculate the limited mean of a matrix of given distributions (positive part only), by dimension and by component.
}
\keyword{internal}
