% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictMean.R
\name{DimCompPosYMean}
\alias{DimCompPosYMean}
\title{A function to calculate the mean of a matrix of given distributions (positive part only), by dimension and by component.}
\usage{
DimCompPosYMean(comp.dist, params.list)
}
\arguments{
\item{comp.dist}{A d*g matrix of strings, describing component distributions by dimension and by component.}

\item{paramas.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A matrix of mean values by dimension and by component.
}
\description{
A function to calculate the mean of a matrix of given distributions (positive part only), by dimension and by component.
}
\keyword{internal}
