% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictVaRCTE.R
\name{DimCompYCDF}
\alias{DimCompYCDF}
\title{A function to calculate the cdf of a matrix of given distributions (with zero inflation), by dimension and by component.}
\usage{
DimCompYCDF(y, comp.dist, zero.prob, params.list)
}
\arguments{
\item{y}{A numeric vector by dimension.}

\item{comp.dist}{A d*g matrix of strings, describing component distributions by dimension and by component.}

\item{zero.prob}{A d*g matrix of numbers between 0 and 1, describing zero probability masses by dimension and by component.}

\item{paramas.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A matrix of cdf values by dimension and by component.
}
\description{
A function to calculate the cdf of a matrix of given distributions (with zero inflation), by dimension and by component.
}
\keyword{internal}
