% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMEzkz.R
\name{EMEzkz}
\alias{EMEzkz}
\title{ECM: E-Step for \code{z.e.obs}, \code{k.e} and \code{z.e.lat}.}
\usage{
EMEzkz(gate.ll, expert.list, ll.list)
}
\arguments{
\item{gate.ll}{An object returned by \code{\link{GateLogit}}.}

\item{expert.list}{An object returned by \code{\link{DimCompExpertLL}}.}

\item{ll.list}{An object returned by \code{\link{GateExpertLL}}.}
}
\value{
\code{z.e.obs},\code{k.e},\code{z.e.lat} Numerical vectors of length N.
}
\description{
ECM: E-Step for \code{z.e.obs}, \code{k.e} and \code{z.e.lat}.
}
\keyword{internal}
