% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMBinomial.R
\name{EMMBinomial}
\alias{EMMBinomial}
\title{ECM: M-Step for Binomial expert.}
\usage{
EMMBinomial(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Binomial expert.
}
\keyword{internal}
