% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMBurr.R
\name{EMMBurr}
\alias{EMMBurr}
\title{ECM: M-Step for Inverse Gaussian expert.}
\usage{
EMMBurr(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Inverse Gaussian expert.
}
\seealso{
\code{\link{comp.kj.params.m.recur}}
}
\keyword{internal}
