% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosEM.R
\name{EMMCompParams}
\alias{EMMCompParams}
\title{ECM: M-Step for updating parameters of positive part of component experts.}
\usage{
EMMCompParams(
  comp.kj.dist,
  comp.kj.params.old,
  tl.k,
  yl.k,
  yu.k,
  tu.k,
  comp.kj.pos.expert.ll,
  comp.kj.pos.expert.tn,
  comp.kj.pos.expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params.kj
)
}
\arguments{
\item{comp.kj.dist}{A string which indicates component distribution.}

\item{comp.kj.params.old}{A vector of numerics: old parameter values for \code{comp.kj.dist}.}

\item{tl.k, yl.k, yu.k, tu.k}{Numerical vectors of length N.}

\item{comp.kj.pos.expert.ll, comp.kj.pos.expert.tn, comp.kj.pos.expert.tn.bar}{See \code{\link{PosExpertLL}}.}

\item{z.e.obs, z.e.lat, k.e}{A numerical vector returned by \code{\link{EMEzkz}}.}

\item{penalty, hyper.params.kj}{See \code{\link{GateExpertLL}}.}
}
\value{
Updated parameter values.
}
\description{
ECM: M-Step for updating parameters of positive part of component experts.
}
\keyword{internal}
