% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMGamma.R
\name{EMMGamma}
\alias{EMMGamma}
\title{ECM: M-Step for Gamma expert.}
\usage{
EMMGamma(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Gamma expert.
}
\keyword{internal}
