% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMGammaCount.R
\name{EMMGammaCount}
\alias{EMMGammaCount}
\title{ECM: M-Step for Gamma Count expert.}
\usage{
EMMGammaCount(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Gamma Count expert.
}
\keyword{internal}
