% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMInverseGaussian.R
\name{EMMInverseGaussian}
\alias{EMMInverseGaussian}
\title{ECM: M-Step for Inverse Gaussian expert.}
\usage{
EMMInverseGaussian(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Inverse Gaussian expert.
}
\keyword{internal}
