% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLognormal.R
\name{EMMLognormal}
\alias{EMMLognormal}
\title{ECM: M-Step for LognOrmal expert.}
\usage{
EMMLognormal(
  lnorm.params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for LognOrmal expert.
}
\keyword{internal}
