% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMNegativeBinomial.R
\name{EMMNegativeBinomial}
\alias{EMMNegativeBinomial}
\title{ECM: M-Step for Negative Binomial expert.}
\usage{
EMMNegativeBinomial(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Negative Binomial expert.
}
\keyword{internal}
