% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMPoisson.R
\name{EMMPoisson}
\alias{EMMPoisson}
\title{ECM: M-Step for Poisson expert.}
\usage{
EMMPoisson(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Poisson expert.
}
\keyword{internal}
