% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMWeibull.R
\name{EMMWeibull}
\alias{EMMWeibull}
\title{ECM: M-Step for Weibull expert.}
\usage{
EMMWeibull(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Weibull expert.
}
\keyword{internal}
