% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMzero.R
\name{EMMzero}
\alias{EMMzero}
\title{ECM: M-Step for \code{zero.prob}.}
\usage{
EMMzero(z.zero.e.obs, z.pos.e.obs, z.zero.e.lat, z.pos.e.lat, k.e)
}
\arguments{
\item{z.zero.e.obs}{An object returned by \code{\link{EMEzzeroobs}}.}

\item{z.pos.e.obs}{An object returned by \code{\link{EMEzzeroobs}}.}

\item{z.zero.e.lat}{An object returned by \code{\link{EMEzzerolat}}.}

\item{z.pos.e.lat}{An object returned by \code{\link{EMEzzerolat}}.}

\item{k.e}{An object returned by \code{\link{EMEzkz}}.}
}
\value{
\code{zero.prob} Updated zero.prob.
}
\description{
ECM: M-Step for \code{zero.prob}.
}
\keyword{internal}
