% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMZTPoisson.R
\name{EMMZTPoisson}
\alias{EMMZTPoisson}
\title{ECM: M-Step for Zero-Truncated Poisson expert.}
\usage{
EMMZTPoisson(
  params.old,
  tl,
  yl,
  yu,
  tu,
  expert.ll,
  expert.tn,
  expert.tn.bar,
  z.e.obs,
  z.e.lat,
  k.e,
  penalty,
  hyper.params
)
}
\description{
ECM: M-Step for Zero-Truncated Poisson expert.
}
\keyword{internal}
