% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertFunction.R
\name{ExpertFunction}
\alias{ExpertFunction}
\title{Expert Function Classes}
\description{
This is the base class for Expert Functions. Users are expected to interact with \code{ExpertFunction} to perform all the computations.
}
\examples{
# p = ExpertFunction$new("poisson")
# p$continuous
# p$distribution
# Estimate the start parameters and get them
# p$initialize_params(c(1,12,2,3,4,6,7,7))
# p$get_params()
# Initialize the penalty using default penalty parameters and get them
# p$initialize_penalty()
# p$get_penalty()
# Without setiing parameters, the object will use preset parameters
# p$get_penalty_params()
# Exposurize function, this will not change the parameters of the ExpertFunction
# p$exposurize(2)
# Calculate the loglikelihood by two methods
# p$ll_exact(5)
# p$ll_not_exact(1,1,2,2)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{distribution}}{(\code{character})\cr
The distribution name specified by the user.}

\item{\code{continuous}}{(\code{logical})\cr
Given distribution is continuous or not, default set to be \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExpertFunction$new()}}
\item \href{#method-initialize_penalty}{\code{ExpertFunction$initialize_penalty()}}
\item \href{#method-initialize_params}{\code{ExpertFunction$initialize_params()}}
\item \href{#method-set_params}{\code{ExpertFunction$set_params()}}
\item \href{#method-get_params}{\code{ExpertFunction$get_params()}}
\item \href{#method-set_penalty_params}{\code{ExpertFunction$set_penalty_params()}}
\item \href{#method-get_penalty_params}{\code{ExpertFunction$get_penalty_params()}}
\item \href{#method-get_penalty}{\code{ExpertFunction$get_penalty()}}
\item \href{#method-get_mean}{\code{ExpertFunction$get_mean()}}
\item \href{#method-get_variance}{\code{ExpertFunction$get_variance()}}
\item \href{#method-get_cdf}{\code{ExpertFunction$get_cdf()}}
\item \href{#method-get_logcdf}{\code{ExpertFunction$get_logcdf()}}
\item \href{#method-get_pdf}{\code{ExpertFunction$get_pdf()}}
\item \href{#method-get_logpdf}{\code{ExpertFunction$get_logpdf()}}
\item \href{#method-get_quantile}{\code{ExpertFunction$get_quantile()}}
\item \href{#method-get_lev}{\code{ExpertFunction$get_lev()}}
\item \href{#method-get_excess}{\code{ExpertFunction$get_excess()}}
\item \href{#method-exposurize}{\code{ExpertFunction$exposurize()}}
\item \href{#method-simulate}{\code{ExpertFunction$simulate()}}
\item \href{#method-ll_exact}{\code{ExpertFunction$ll_exact()}}
\item \href{#method-ll_not_exact}{\code{ExpertFunction$ll_not_exact()}}
\item \href{#method-EM_notexact}{\code{ExpertFunction$EM_notexact()}}
\item \href{#method-EM_exact}{\code{ExpertFunction$EM_exact()}}
\item \href{#method-clone}{\code{ExpertFunction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ExpertFunction Class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$new(distribution, params = list(), penalty_params = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distribution}}{(\code{character})\cr
The distribution name specified by the user.}

\item{\code{params}}{(\code{list})\cr
The parameters for the given distribution.
Could be preset by users or postset by running \code{ExpertFunction$params_init(observations)}}

\item{\code{penalty_params}}{(\code{list})\cr
The penalty parameters used to determine the penalty applied on likelihood, will be further used in (\code{initialize_penalty}) methods
Calculate the penalty value of the Expert Distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initialize_penalty"></a>}}
\if{latex}{\out{\hypertarget{method-initialize_penalty}{}}}
\subsection{Method \code{initialize_penalty()}}{
Prerequisite: self$params and self$penalty_params need to be set. Otherwise will return the default penalty value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$initialize_penalty()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{default}}{(\code{logical})\cr
Use default penalty parameters if default is TRUE, else use user defined penalty parameters. Default set to be TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initialize_params"></a>}}
\if{latex}{\out{\hypertarget{method-initialize_params}{}}}
\subsection{Method \code{initialize_params()}}{
Estimate all the parameters for this expert function based on observation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$initialize_params(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{(\code{numeric})\cr
The observations from the preset expert distribution
Set the parameters for the expert function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$set_params(expert_params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_params}}{(\code{logical})\cr
The parameters for the given distribution.
Get the parameters for the expert function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-get_params}{}}}
\subsection{Method \code{get_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_params()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{(\code{logical})\cr
The parameters for the given distribution.
Set the penalty parameters
If there are no expert_penalty_params given, use default penalty parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_penalty_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_penalty_params}{}}}
\subsection{Method \code{set_penalty_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$set_penalty_params(expert_penalty_params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_penalty_params}}{(\code{logical})\cr
The parameters for the given distribution. Could be preset by users or postset by running \code{ExpertFunction$initialize_params(observations)}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_penalty_params"></a>}}
\if{latex}{\out{\hypertarget{method-get_penalty_params}{}}}
\subsection{Method \code{get_penalty_params()}}{
Get the penalty parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_penalty_params()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_penalty"></a>}}
\if{latex}{\out{\hypertarget{method-get_penalty}{}}}
\subsection{Method \code{get_penalty()}}{
Get the penalty value, penalty value cannot be set. It can only been initialized.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_penalty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mean"></a>}}
\if{latex}{\out{\hypertarget{method-get_mean}{}}}
\subsection{Method \code{get_mean()}}{
Get mean of this expert function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_mean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variance"></a>}}
\if{latex}{\out{\hypertarget{method-get_variance}{}}}
\subsection{Method \code{get_variance()}}{
Get variance of this expert function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cdf"></a>}}
\if{latex}{\out{\hypertarget{method-get_cdf}{}}}
\subsection{Method \code{get_cdf()}}{
Get the cdf of this expert function given RV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_cdf(q)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{The value of random variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_logcdf"></a>}}
\if{latex}{\out{\hypertarget{method-get_logcdf}{}}}
\subsection{Method \code{get_logcdf()}}{
Get the logcdf of this expert function given RV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_logcdf(q)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{The value of random variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pdf"></a>}}
\if{latex}{\out{\hypertarget{method-get_pdf}{}}}
\subsection{Method \code{get_pdf()}}{
Get the pdf of this expert function given RV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_pdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value of random variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_logpdf"></a>}}
\if{latex}{\out{\hypertarget{method-get_logpdf}{}}}
\subsection{Method \code{get_logpdf()}}{
Get the logpdf of this expert function given RV
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_logpdf(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value of random variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_quantile"></a>}}
\if{latex}{\out{\hypertarget{method-get_quantile}{}}}
\subsection{Method \code{get_quantile()}}{
Get the quantile of this expert function given probability
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_quantile(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{The probability}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_lev"></a>}}
\if{latex}{\out{\hypertarget{method-get_lev}{}}}
\subsection{Method \code{get_lev()}}{
Get the Limited Expected Value of the expert function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_lev(u)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{u}}{(\code{numeric})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_excess"></a>}}
\if{latex}{\out{\hypertarget{method-get_excess}{}}}
\subsection{Method \code{get_excess()}}{
Get the Excess of the function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$get_excess(u)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{u}}{(\code{numeric})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exposurize"></a>}}
\if{latex}{\out{\hypertarget{method-exposurize}{}}}
\subsection{Method \code{exposurize()}}{
Exposurize the parameters, return a copy of the object with the updated parameters
Note this function will not change the parameters of the original object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$exposurize(exposure = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure}}{(\code{numeric})\cr
The exposure applied to the parameters, default set to be 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-simulate}{}}}
\subsection{Method \code{simulate()}}{
Simulate value based on distribution parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$simulate(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{numeric})\cr
The number of simulated values you want
Calculate the exact log likelihood of the Expert Distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ll_exact"></a>}}
\if{latex}{\out{\hypertarget{method-ll_exact}{}}}
\subsection{Method \code{ll_exact()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$ll_exact(y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{(\code{numeric})\cr
The observations from the preset expert distribution
Calculate the non-exact log likelihood of the Expert Distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ll_not_exact"></a>}}
\if{latex}{\out{\hypertarget{method-ll_not_exact}{}}}
\subsection{Method \code{ll_not_exact()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$ll_not_exact(tl, yl, yu, tu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tl}}{(\code{numeric})\cr A vector of length N: lower bounds of truncation.}

\item{\code{yl}}{(\code{numeric})\cr A vector of length N: lower bounds of censoring.}

\item{\code{yu}}{(\code{numeric})\cr A vector of length N: upper bounds of censoring.}

\item{\code{tu}}{(\code{numeric})\cr A vector of length N: upper bounds of truncation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EM_notexact"></a>}}
\if{latex}{\out{\hypertarget{method-EM_notexact}{}}}
\subsection{Method \code{EM_notexact()}}{
Perform the EM optimization with non-exact observations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$EM_notexact(
  expert_old,
  tl,
  yl,
  yu,
  tu,
  exposure,
  z_e_obs,
  z_e_lat,
  k_e,
  penalty,
  pen_params
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_old}}{Old expert function}

\item{\code{tl}}{(\code{numeric})\cr A vector of length N: lower bounds of truncation.}

\item{\code{yl}}{(\code{numeric})\cr A vector of length N: lower bounds of censoring.}

\item{\code{yu}}{(\code{numeric})\cr A vector of length N: upper bounds of censoring.}

\item{\code{tu}}{(\code{numeric})\cr A vector of length N: upper bounds of truncation.}

\item{\code{exposure}}{A vector of length N: exposures for observations}

\item{\code{z_e_obs}}{Calculated from E-step}

\item{\code{z_e_lat}}{Calculated from E-step}

\item{\code{k_e}}{Calculated from E-step}

\item{\code{penalty}}{T/F: whether penalty is imposed}

\item{\code{pen_params}}{A vector of penalty parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EM_exact"></a>}}
\if{latex}{\out{\hypertarget{method-EM_exact}{}}}
\subsection{Method \code{EM_exact()}}{
Perform the EM optimization with exact observations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$EM_exact(expert_old, ye, exposure, z_e_obs, penalty, pen_params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_old}}{Old expert function}

\item{\code{ye}}{A vector of length N: exact observations}

\item{\code{exposure}}{A vector of length N: exposures for observations}

\item{\code{z_e_obs}}{Calculated from E-step}

\item{\code{penalty}}{T/F: whether penalty is imposed}

\item{\code{pen_params}}{A vector of penalty parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertFunction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
