% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertGammaCount.R
\name{ExpertGammaCount}
\alias{ExpertGammaCount}
\title{Expert Function: Gamma Count.}
\usage{
ExpertGammaCount(tl, yl, yu, tu, m, s)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{m}{A vector of length \code{1}: Gamma Count shape parameters.}

\item{s}{A vector of length \code{1}: Gamma Count dispersion parameters.}
}
\value{
A list of matrices of expert loglikelihood for Gamma Count.
}
\description{
Expert Function: Gamma Count.
}
\seealso{
\code{\link[rmutil]{GammaCount}}.
}
\keyword{internal}
