% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertInverseGaussian.R
\name{ExpertInvgauss}
\alias{ExpertInvgauss}
\title{Expert Function: Inverse Gaussian.}
\usage{
ExpertInvgauss(tl, yl, yu, tu, mean.mu, shape.lambda)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{mean.mu}{A vector of length \code{1}: Inverse Gaussian mean parameters.}

\item{shape.lambda}{A vector of length \code{1}: Inverse Gaussian shape parameters.}
}
\value{
A list of matrices of expert loglikelihood for Inverse Gaussian.
}
\description{
Expert Function: Inverse Gaussian.
}
\seealso{
\code{\link[statmod]{invgauss}}.
}
\keyword{internal}
