% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertLognormal.R
\name{ExpertLognormal}
\alias{ExpertLognormal}
\title{Expert Function: Log Normal.}
\usage{
ExpertLognormal(tl, yl, yu, tu, meanlog, sdlog)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{meanlog}{A vector of length \code{1}: Lognormal mean parameters.}

\item{sdlog}{A vector of length \code{1}: Lognormal sd parameters.}
}
\value{
A list of matrices of expert loglikelihood for Log Normal.
}
\description{
Expert Function: Log Normal.
}
\seealso{
\code{\link[stats]{Lognormal}}.
}
\keyword{internal}
