% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertLibrary.R, R/ExpertMatrix.R
\name{ExpertMatrix}
\alias{ExpertMatrix}
\alias{ExpertLibrary}
\title{Expert Library Classes}
\description{
This is the library of all the Expert Functions. In another word, it stored all the metadata of the Expert Functions.
This object is a singleton, which means it should be only created when the package is loaded.
It is also hidden in the global environment. Please do not clear the hidden objects
And it you cleared the hidden objects. Please reload the package to fix that.

This is the matrix class of Expert Functions. Users are expected to perform all vector operations through ExpertMatrix to reduce index errors.
}
\examples{
params = matrix(list( list(meanlog = 1, sdlog = 2),  list(meanlog = 1, sdlog  = 2),
                      list(meanlog = 3, sdlog = 1),  list(meanlog = 5, sdlog = 1)), nrow = 2)
penalty_params = matrix( list( c(2,1,2), c(1,1,1), c(3,1,2), c(1,3,4)), nrow = 2)
expert_names = matrix( c("lognormal", "lognormal", "lognormal", "lognormal"), nrow = 2)
# ExpertMatrixObj = ExpertMatrix$new(expert_matrix = expert_names, expert_params_matrix = params)
# ExpertMatrixObj$set_penalty_params(penalty_params)
# ExpertMatrixObj$set_params(params)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{continuous_experts}}{(\code{NULL})\cr
The names of continuous ExpertFunction}

\item{\code{discrete_experts}}{(\code{NULL})\cr
The names of discrete ExpertFunction}

\item{\code{distribution_names}}{(\code{NULL})\cr
The names of all expert functions
Register a new Expert Function}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-register}{\code{ExpertLibrary$register()}}
\item \href{#method-is_continous}{\code{ExpertLibrary$is_continous()}}
\item \href{#method-is_discrete}{\code{ExpertLibrary$is_discrete()}}
\item \href{#method-is_valid_expert}{\code{ExpertLibrary$is_valid_expert()}}
\item \href{#method-clone}{\code{ExpertLibrary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-register"></a>}}
\if{latex}{\out{\hypertarget{method-register}{}}}
\subsection{Method \code{register()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertLibrary$register(expert_name, continuous)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_name}}{(\code{character})\cr
The name of the expert function}

\item{\code{continuous}}{(\code{logical})\cr
Is this new expert function continuous?
Is given Expert Function continuous?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_continous"></a>}}
\if{latex}{\out{\hypertarget{method-is_continous}{}}}
\subsection{Method \code{is_continous()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertLibrary$is_continous(expert_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_name}}{(\code{character})\cr
The name of the expert function
Is given Expert Function discrete?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_discrete"></a>}}
\if{latex}{\out{\hypertarget{method-is_discrete}{}}}
\subsection{Method \code{is_discrete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertLibrary$is_discrete(expert_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_name}}{(\code{character})\cr
The name of the expert function
Is given Expert Function got a correct name?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_valid_expert"></a>}}
\if{latex}{\out{\hypertarget{method-is_valid_expert}{}}}
\subsection{Method \code{is_valid_expert()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertLibrary$is_valid_expert(expert_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_name}}{(\code{character})\cr
The name of the expert function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertLibrary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nrow}}{(\code{numeric})\cr
The number of rows of ExpertMatrix}

\item{\code{ncol}}{(\code{numeric})\cr
The number of columns of ExpertMatrix}

\item{\code{expert_matrix}}{(\code{matrix})\cr
The matrix object to store all the ExpertFunction R6 objects}

\item{\code{penalty_params}}{(\code{matrix})\cr
The matrix object to store all the penalty parameters for the Expert Functions}

\item{\code{expert_params}}{(\code{matrix})\cr
The matrix object to store all the parameters for the Expert Functions
Initialze a Expert Matrix Class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExpertMatrix$new()}}
\item \href{#method-select}{\code{ExpertMatrix$select()}}
\item \href{#method-exposurize}{\code{ExpertMatrix$exposurize()}}
\item \href{#method-get_mean}{\code{ExpertMatrix$get_mean()}}
\item \href{#method-get_variance}{\code{ExpertMatrix$get_variance()}}
\item \href{#method-set_penalty_params}{\code{ExpertMatrix$set_penalty_params()}}
\item \href{#method-set_params}{\code{ExpertMatrix$set_params()}}
\item \href{#method-get_penalty_value}{\code{ExpertMatrix$get_penalty_value()}}
\item \href{#method-count_params}{\code{ExpertMatrix$count_params()}}
\item \href{#method-clone}{\code{ExpertMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$new(expert_matrix, expert_params_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_matrix}}{(\code{matrix})\cr
The name of Expert Functions, stored as a matrix of characters.}

\item{\code{expert_params_matrix}}{(\code{matrix})\cr
The parameters of Expert Functions, stored as a matrix of lists.
select the expert function object inside expert matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select"></a>}}
\if{latex}{\out{\hypertarget{method-select}{}}}
\subsection{Method \code{select()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$select(row_index = 0, col_index = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col_index}}{Exposurize all the expert functions inside the expert matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exposurize"></a>}}
\if{latex}{\out{\hypertarget{method-exposurize}{}}}
\subsection{Method \code{exposurize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$exposurize(exposure)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exposure}}{The exposurized value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mean"></a>}}
\if{latex}{\out{\hypertarget{method-get_mean}{}}}
\subsection{Method \code{get_mean()}}{
Return the mean matrix of the ExpertMatrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$get_mean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_variance"></a>}}
\if{latex}{\out{\hypertarget{method-get_variance}{}}}
\subsection{Method \code{get_variance()}}{
Return the variance matrix of the ExpertMatrix
set the penalty parameters for all the expert functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$get_variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_penalty_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_penalty_params}{}}}
\subsection{Method \code{set_penalty_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$set_penalty_params(expert_penalty_params_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_penalty_params_matrix}}{(\code{matrix})
To be more specific, it is a matrix of list like this

penalty_params = matrix( list( c(2,1,2), c(1,1,1), c(3,1,2), c(1,3,4)), nrow = 2)
set the parameters for all the expert functions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$set_params(expert_params_matrix)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expert_params_matrix}}{(\code{matrix})
To be more specific, it is a matrix of list of named list like this

params = matrix(list( list(meanlog = 1, sdlog = 2),  list(meanlog = 1, sdlog  = 2),
list(meanlog = 3, sdlog = 1),  list(meanlog = 5, sdlog = 1)), nrow = 2)
Get the sum of penalty value of all the expert functions. Users are required to set the penalty parameters before call this function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_penalty_value"></a>}}
\if{latex}{\out{\hypertarget{method-get_penalty_value}{}}}
\subsection{Method \code{get_penalty_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$get_penalty_value(penalty)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
value
The total penalty value of Expert Matrix
Count the total number of parameters (for the Expert Functions) that exist in the Expert Matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-count_params"></a>}}
\if{latex}{\out{\hypertarget{method-count_params}{}}}
\subsection{Method \code{count_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$count_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
result (\code{integer})
Total number of parameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpertMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
