% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertNegativeBinomial.R
\name{ExpertNbinom}
\alias{ExpertNbinom}
\title{Expert Function: Negative Binomial.}
\usage{
ExpertNbinom(tl, yl, yu, tu, size.n, prob.p)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{size.n}{A vector of length \code{1}: Negative Binomial size.n parameters.}

\item{prob.p}{A vector of length \code{1}: Negative Binomial prob.p parameters.}
}
\value{
A list of matrices of expert loglikelihood for Negative Binomial.
}
\description{
Expert Function: Negative Binomial.
}
\seealso{
\code{\link[stats]{NegBinomial}}.
}
\keyword{internal}
