% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertPoisson.R
\name{ExpertPoisson}
\alias{ExpertPoisson}
\title{Expert Function: Poisson.}
\usage{
ExpertPoisson(tl, yl, yu, tu, mean.theta)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{mean.theta}{A vector of length \code{1}: Poisson mean parameters.}
}
\value{
A list of matrices of expert loglikelihood for Poisson.
}
\description{
Expert Function: Poisson.
}
\seealso{
\code{\link[stats]{Poisson}}.
}
\keyword{internal}
