% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpertWeibull.R
\name{ExpertWeibull}
\alias{ExpertWeibull}
\title{Expert Function: Weibull.}
\usage{
ExpertWeibull(tl, yl, yu, tu, shape.k, scale.lambda)
}
\arguments{
\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{shape.k}{A vector of length \code{1}: Weibull shape parameters.}

\item{scale.lambda}{A vector of length \code{1}: Weibull scale parameters.}
}
\value{
A list of matrices of expert loglikelihood for Weibull.
}
\description{
Expert Function: Weibull.
}
\seealso{
\code{\link[stats]{Weibull}}.
}
\keyword{internal}
