% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitInterface.R
\name{FitLRMoE}
\alias{FitLRMoE}
\title{Main fitting function of LRMoE}
\usage{
FitLRMoE(
  Y,
  X,
  alpha_init,
  comp_dist,
  params_list,
  exposure = NULL,
  exact_Y = FALSE,
  penalty = TRUE,
  pen_alpha = 5,
  pen_params = NULL,
  eps = 0.001,
  alpha_iter_max = 3,
  ecm_iter_max = 200,
  grad_jump = TRUE,
  grad_seq = NULL,
  print_steps = TRUE
)
}
\arguments{
\item{Y}{A N by d (\code{exact_Y=T}) or N by 4d (\code{exact_Y=F}) matrix of numerics,
where N is sample size and d is the dimension of each obsevation.
If the size is N by 4d, Each block of four columns should be organized as \code{(tl, yl, yu, tu)}, representing the
truncation lower bound, censoring lower bound, censoring upper bound and truncation upper bound.}

\item{X}{A N*P matrix of numerics, where P is the number of covariates.
The first column of \code{X} should be 1, which is the intercept.}

\item{alpha_init}{A g*P matrix of numerics, which contains initial guess of the logit regression coefficients.
The last row should all be zero, representing the default latent class.
If no initialization is provided, all coefficients are set to zero.}

\item{comp_dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{params_list}{A d * g matrix of list with paramster names and values,
which is the initial parameter guess for the corresponding \code{comp_dist}.}

\item{exposure}{A vector of length N, representing the exposure of the observations (how long it has been observed).}

\item{exact_Y}{TRUE/FALSE: whether \code{Y} is observed exactly, or with censoring and/or truncation.}

\item{penalty}{TRUE/FALSE: whether the parameters are penalized for their magnitude.
Default (and recommended) is TRUE.}

\item{pen_alpha}{A numeric, which contains penalties for \code{alpha_init}.
If \code{penalty=T} but no \code{pen_alpha} is provided, a constant is used.}

\item{pen_params}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the corresponding penalty for \code{params.init}.}

\item{eps}{Stopping criteria for loglikelihood convergence. Default is \code{1e-03}.}

\item{alpha_iter_max}{Maximum number of iterations for updating alpha. Defauls is 5.}

\item{ecm_iter_max}{Maximum number of iterations for ECM. Default is 200.}

\item{grad_jump}{TRUE/FALSE: whether to use an approximated gradient jump to speed up convergence.}

\item{grad_seq}{How are the gradient sequence selected. Default is \code{2^(seq(8)-1)-1}.}

\item{print_steps}{TRUE/FALSE: whether paramater updates are printed on screen. Default is TRUE.}
}
\description{
Main fitting function of LRMoE
}
