% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loglik.R
\name{GateExpertLL}
\alias{GateExpertLL}
\title{Computes the entire loglikelihood, including gating function and experts.}
\usage{
GateExpertLL(
  alpha,
  gate.ll,
  expert.list,
  penalty = TRUE,
  hyper.alpha = NULL,
  hyper.params = NULL
)
}
\arguments{
\item{alpha}{A \code{g*P} matrix, where \code{g} is the number of components and P is the number of covariates.}

\item{gate.ll}{An object returned by function \code{\link{GateLogit}}.}

\item{expert.list}{An object returned by function \code{\link{DimCompExpertLL}}.}

\item{penalty}{TRUE/FALSE, which indicates whether parameter penalty should be applied. Default (and recommended) is TRUE.}

\item{hyper.alpha}{A numeric, which penalizes the magnitude of \code{alpha}.}

\item{hyper.params}{A list of length \code{d}. Each element is a sublist of length \code{g}.
Each element of a sublist is a vector of numerics, which penalizes expert parameters. See also \code{\link{DimCompExpertPenalty}}.}
}
\description{
Computes the entire loglikelihood, including gating function and experts.
}
\keyword{internal}
