% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{GateLogit}
\alias{GateLogit}
\title{Computes the logit regression weights in log.}
\usage{
GateLogit(x, alpha)
}
\arguments{
\item{x}{An N * P covariate matrix, where N is sample size. The first column MUST be 1.}

\item{alpha}{A g * P matrix. Logit regression coefficients.}
}
\value{
The log of \code{tcrossprod(x,alpha)} with dim N * g, with each row normalized by \code{rowLogSumExps}
}
\description{
Computes the logit regression weights in log.
}
\keyword{internal}
