% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictVaRCTE.R
\name{indCDFPrior}
\alias{indCDFPrior}
\title{Calculate the cdf of y, given a fixed covariate vector X and a model.}
\usage{
indCDFPrior(y, X, alpha, comp.dist, zero.prob, params.list)
}
\arguments{
\item{y}{A numeric vector by dimension.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A vector of cdf values by dimension.
}
\description{
Calculate the cdf of y, given a fixed covariate vector X and a model.
}
\keyword{internal}
