% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictVaRCTE.R
\name{indSolveQuantilePrior}
\alias{indSolveQuantilePrior}
\title{Predict the VaR of y, given a fixed covariate vector X and a model.}
\usage{
indSolveQuantilePrior(X, alpha, comp.dist, zero.prob, params.list, prob = NULL)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}

\item{prob}{A vector of probabilities. Default is a vector of length d of 0.95, if no value is provided.}
}
\value{
A vector of VaR by dimension.
}
\description{
Predict the VaR of y, given a fixed covariate vector X and a model.
}
\keyword{internal}
