% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{LogLikelihoodExact}
\alias{LogLikelihoodExact}
\title{Calculate the log likelihood}
\usage{
LogLikelihoodExact(Y, gate, model, exposure)
}
\arguments{
\item{Y}{(\code{matrix})\cr
The numerical observation matrix with IxJ dimentions}

\item{gate}{(\code{matrix})\cr
The numerical gate matrix with IxD dimentions}

\item{model}{(\code{ExpertMatrix})\cr
The JxD ExpertMatrix S6 Class which contained ExpertFunctions}

\item{exposure}{(\code{matrix})\cr
The Ix1 numerical vector which specify the exposure for each observation.}
}
\description{
Calculate the log likelihood based on the observations, gate matrix, model matrix and exposure matrix.
}
\keyword{internal}
