% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{LogLikelihoodNotExact}
\alias{LogLikelihoodNotExact}
\title{Calculate the log likelihood}
\usage{
LogLikelihoodNotExact(Y, gate, model, exposure)
}
\arguments{
\item{Y}{(\code{array})\cr
The observation matrix with Ix4J dimentions, where each entry is a tuple of this format (tl, yl, tu, yu).
Keep in mind it means there will be three dimention for this matrix, and the last dim need to be 4.}

\item{gate}{(\code{matrix})\cr
The numerical gate matrix with IxD dimentions}

\item{model}{(\code{ExpertMatrix})\cr
The JxD ExpertMatrix S6 Class which contained ExpertFunctions}

\item{exposure}{(\code{matrix})\cr
The Ix1 numerical vector which specify the exposure for each observation.}
}
\description{
Calculate the log likelihood based on the observations, gate matrix, model matrix and exposure matrix.
Keep in mind our observation matrix now have a different format
}
