% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{LRMoE-package}
\alias{LRMoE}
\alias{LRMoE-package}
\title{LRMoE: Logit-Weighted Reduced Mixture-of-Experts}
\description{
The Logit-Weighted Reduced Mixture-of-Experts (LRMoE) 
  proposed by Fung et al. (2019) is a flexible framework actuarial loss modelling.
  For more details, see Fung et al. (2019) "A class of mixture of experts models for general insurance: Theoretical developments"
  published in Insurance: Mathematics and Economics, and
  Fung et al. (2019) "A CLASS OF MIXTURE OF EXPERTS MODELS FOR GENERAL INSURANCE: APPLICATION TO CORRELATED CLAIM FREQUENCIES"
  in ASTIN Bulletin: The Journal of the IAA.
}
