% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRMoEDemoData.R
\docType{data}
\name{LRMoEDemoData}
\alias{LRMoEDemoData}
\alias{X}
\alias{Y}
\alias{X.obs}
\alias{Y.obs}
\title{Demo data for \code{LRMoE}.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10000 rows and 5 columns.

An object of class \code{matrix} (inherits from \code{array}) with 10000 rows and 8 columns.

An object of class \code{matrix} (inherits from \code{array}) with 9828 rows and 5 columns.

An object of class \code{matrix} (inherits from \code{array}) with 9828 rows and 8 columns.
}
\usage{
X

Y

X.obs

Y.obs
}
\description{
\itemize{
  \item \code{X}: A matrix of covariates required by the LRMoE fitting function (complete).
  \item \code{Y}: A matrix of response required by the LRMoE fitting function (complete).
  \item \code{X.obs}: A matrix of covariates required by the LRMoE fitting function (after truncation and censoring).
  \item \code{Y.obs}: A matrix of response required by the LRMoE fitting function (after truncation and censoring).
}
}
\keyword{data}
\keyword{datasets}
