% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRMoEFit.R
\name{LRMoEFit}
\alias{LRMoEFit}
\title{Fit an LRMoE model.}
\usage{
LRMoEFit(
  Y,
  X,
  n.comp = 2,
  comp.dist = NULL,
  alpha.init = NULL,
  zero.init = NULL,
  params.init = NULL,
  penalty = TRUE,
  hyper.alpha = NULL,
  hyper.params = NULL,
  eps = 0.001,
  alpha.iter.max = 5,
  ecm.iter.max = 200,
  grad.jump = TRUE,
  grad.period = 5,
  grad.seq = 2^(seq(8) - 1) - 1,
  print = TRUE
)
}
\arguments{
\item{Y}{A N*4d matrix of numerics, where N is sample size and d is the dimension of each obsevation.
Each block of four columns should be organized as \code{(tl, yl, yu, tu)}, representing the
truncation lower bound, censoring lower bound, censoring upper bound and truncation upper bound.}

\item{X}{A N*P matrix of numerics, where P is the number of covariates.
The first column of \code{X} should be 1, which is the intercept.}

\item{n.comp}{A numeric which indicates the number of experts desired to fit the data. Default value is 2.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.
See below for more details.}

\item{alpha.init}{A g*P matrix of numerics, which contains initial guess of the logit regression coefficients.
The last row should all be zero, representing the default latent class.
If no initialization is provided, all coefficients are set to zero.}

\item{zero.init}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.
If the corresponding entry in \code{comp.dist} is not zero-inflated, zero value must by supplied.}

\item{params.init}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the initial parameter guess for the corresponding \code{comp.dist}.}

\item{penalty}{TRUE/FALSE: whether the parameters are penalized for their magnitude.
Default (and recommended) is TRUE.}

\item{hyper.alpha}{A numeric, which contains penalties for \code{alpha.init}.
If \code{penalty=T} but no \code{hyper.alpha} is provided, a constant is used.}

\item{hyper.params}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the corresponding penalty for \code{params.init}.}

\item{eps}{Stopping criteria for loglikelihood convergence. Default is \code{1e-03}.}

\item{alpha.iter.max}{Maximum number of iterations for updating alpha. Defauls is 5.}

\item{ecm.iter.max}{Maximum number of iterations for ECM. Default is 200.}

\item{grad.jump}{TRUE/FALSE: whether to use an approximated gradient jump to speed up convergence.}

\item{grad.period}{How often should \code{grad.jump} occur. Default is every 5 iterations.}

\item{grad.seq}{How are the gradient sequence selected. Default is \code{2^(seq(8)-1)-1}.}

\item{print}{TRUE/FALSE: whether paramater updates are printed on screen. Default is TRUE.}
}
\description{
Fit an LRMoE model.
}
