% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaCountUtils.R
\name{pgammacount}
\alias{pgammacount}
\title{Modified \code{\link[rmutil]{GammaCount}} cdf for better numerical accuracy.}
\usage{
pgammacount(q, m, s, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{m, s}{Paramaters of Gamma count distribution.}

\item{log.p}{TRUE/FALSE: whether log.p should be returned.}
}
\description{
Modified \code{\link[rmutil]{GammaCount}} cdf for better numerical accuracy.
}
\seealso{
\code{\link[rmutil]{GammaCount}}.
}
