% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotClassPrior.R
\name{PlotClassPrior}
\alias{PlotClassPrior}
\title{Plots a stacked bar chart of latent class probabilities, given a covariate vector.}
\usage{
PlotClassPrior(X, alpha, title = "Latent Class Probabilities")
}
\arguments{
\item{X}{A vector of covariates for one policyholder.}

\item{alpha}{A matrix of logit regression coeffiecients.}

\item{title}{A text string for plot title.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots a stacked bar chart of latent class probabilities, given a covariate vector.
}
\seealso{
\code{\link{PredictClassPrior}}
}
