% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPropPosterior.R
\name{PlotPropPosterior}
\alias{PlotPropPosterior}
\title{Plots two stacked bar charts of most likely latent class proportions, which contrasts prior and posterior latent class proportions, given a covariate matrix.}
\usage{
PlotPropPosterior(
  Y,
  X,
  alpha,
  comp.dist,
  zero.prob,
  params.list,
  title = "Proportion of Latent Classes"
)
}
\arguments{
\item{Y}{A matrix of observed responses for \code{X}.}

\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}

\item{title}{A text string for plot title.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plots two stacked bar charts of most likely latent class proportions, which contrasts prior and posterior latent class proportions, given a covariate matrix.
}
\seealso{
\code{\link{PredictClassPosterior}}
}
