% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosCountParams.R
\name{PosCountParams}
\alias{PosCountParams}
\title{Count the number of parameters in the positive part.}
\usage{
PosCountParams(comp.dist)
}
\arguments{
\item{comp.dist}{A d*g matrix of strings describing the component distributions.}
}
\value{
The number of parameters in the positive part.
}
\description{
Count the number of parameters in the positive part.
}
\keyword{internal}
