% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosDensity.R
\name{PosDensity}
\alias{PosDensity}
\title{Generate a series of density values, given a model}
\usage{
PosDensity(comp.dist, params, y.series)
}
\arguments{
\item{comp.dist}{A string which indicates the expert function.
\itemize{
    \item \code{gamma}: Gamma
    \item \code{lnorm}: Log Normal
    \item \code{invgauss}: Inverse Gaussian
    \item \code{weibull}: Weibull
    \item \code{burr}: Burr
    \item \code{poisson}: Poisson
    \item \code{nbinom}: Negative Binomial
    \item \code{binom}: Binomial
    \item \code{gammacount}: Gamma Count
    \item \code{ZI-root}: Zero-inflated versions of the distributions above, e.g. \code{ZI-gamma}.
}}

\item{params}{A vector of parameter values.
\itemize{
    \item \code{gamma}: \code{(m, theta)}
    \item \code{lnorm}: \code{(meanlog, sdlog)}
    \item \code{invgauss}: \code{(mean.mu, scale.lambda)}
    \item \code{weibull}: \code{(shape.k, scale.lambda)}
    \item \code{burr}: \code{(shape1.k, shape2.c, scale.lambda)}
    \item \code{poisson}: \code{(mean.theta)}
    \item \code{nbinom}: \code{(size.n, prob.p)}
    \item \code{binom}: \code{(size.n, prob.p)}
    \item \code{gammacount}: \code{(m, s)}
}}

\item{y.series}{A vector of y values.}
}
\value{
A vector of density values evaluatd at y given.
}
\description{
Generate a series of density values, given a model
}
\keyword{internal}
