% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosLoglik.R
\name{PosExpertLL}
\alias{PosExpertLL}
\title{A switch to calculate the loglikelihood of expert functions.}
\usage{
PosExpertLL(ind.dist, tl, yl, yu, tu, params)
}
\arguments{
\item{ind.dist}{A string which indicates the expert function.
\itemize{
    \item \code{gamma}: Gamma
    \item \code{lnorm}: Log Normal
    \item \code{invgauss}: Inverse Gaussian
    \item \code{weibull}: Weibull
    \item \code{burr}: Burr
    \item \code{poisson}: Poisson
    \item \code{ztpoisson}: Zero-Truncated Poisson
    \item \code{nbinom}: Negative Binomial
    \item \code{binom}: Binomial
    \item \code{gammacount}: Gamma Count
    \item \code{ZI-root}: Zero-inflated versions of the distributions above, e.g. \code{ZI-gamma}.
}}

\item{tl}{A vector of length N: lower bounds of truncation.}

\item{yl}{A vector of length N: lower bounds of censoring.}

\item{yu}{A vector of length N: upper bounds of censoring.}

\item{tu}{A vector of length N: upper bounds of truncation.}

\item{params}{A vector of parameters for \code{ind.dist}.}
}
\value{
A list of matrices of expert loglikelihood for the chosen distribution \code{ind.dist}.
\itemize{
    \item \code{expert.ll}: An N*1 matrix of loglikelihood: density for exact observation; probability for censored observation.
    \item \code{expert.tn}: An N*1 matrix of loglikelihood within truncation limits.
    \item \code{expert.tn.bar}: An N*1 matrix of loglikelihood outside truncation limits, which is \code{log(1-exp(expert.tn))}.
}
}
\description{
A switch to calculate the loglikelihood of expert functions.
}
\seealso{
\code{\link{ExpertGamma}}, \code{\link{ExpertLognormal}}, \code{\link{ExpertInvgauss}}, \code{\link{ExpertWeibull}}, \code{\link{ExpertBurr}},
         \code{\link{ExpertPoisson}}, \code{\link{ExpertNbinom}}, \code{\link{ExpertBinom}}, \code{\link{ExpertGammaCount}}
}
\keyword{internal}
