% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosMean.R
\name{PosMean}
\alias{PosMean}
\title{A switch to calculate the mean of a given distribution.}
\usage{
PosMean(comp.dist, params)
}
\arguments{
\item{comp.dist}{A string which indicates the expert function.
\itemize{
    \item \code{gamma}: Gamma
    \item \code{lnorm}: Log Normal
    \item \code{invgauss}: Inverse Gaussian
    \item \code{weibull}: Weibull
    \item \code{burr}: Burr
    \item \code{poisson}: Poisson
    \item \code{ztpoisson}: Zero-Truncated Poisson
    \item \code{nbinom}: Negative Binomial
    \item \code{binom}: Binomial
    \item \code{gammacount}: Gamma Count
    \item \code{ZI-root}: Zero-inflated versions of the distributions above, e.g. \code{ZI-gamma}.
}}

\item{params}{A vector of parameter values.
\itemize{
    \item \code{gamma}: \code{(m, theta)}
    \item \code{lnorm}: \code{(meanlog, sdlog)}
    \item \code{invgauss}: \code{(mean.mu, scale.lambda)}
    \item \code{weibull}: \code{(shape.k, scale.lambda)}
    \item \code{burr}: \code{(shape1.k, shape2.c, scale.lambda)}
    \item \code{poisson}: \code{(mean.theta)}
     \item \code{ztpoisson}: \code{(mean.theta)}
    \item \code{nbinom}: \code{(size.n, prob.p)}
    \item \code{binom}: \code{(size.n, prob.p)}
    \item \code{gammacount}: \code{(m, s)}
}}
}
\description{
A switch to calculate the mean of a given distribution.
}
\keyword{internal}
