% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosName.R
\name{PosName}
\alias{PosName}
\title{A function to format model output: add names for parameters}
\usage{
PosName(comp.dist)
}
\description{
A function to format model output: add names for parameters
}
\seealso{
\code{\link{pos.expert.loglik.calc}}
}
\keyword{internal}
