% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict_class_posterior}
\alias{predict_class_posterior}
\title{predict_class_posterior}
\usage{
predict_class_posterior(
  Y,
  X,
  alpha,
  model,
  exposure_past = list(),
  exact_Y = F
)
}
\arguments{
\item{Y}{A matrix of responses.}

\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{model}{A matrix specifying the expert functions.}

\item{exposure_past}{A vector indicating the time exposure (past) of each observation. If nothing is supplied, it is set to 1.0 by default.}

\item{exact_Y}{\code{true} or \code{false} (default), indicating if \code{Y} is observed exactly or with censoring and truncation.}
}
\value{
prob:
A matrix of latent class probabilities.

max_prob_idx:
A matrix of the most likely latent class for each observation.
}
\description{
: Predicts the latent class probabilities, given observations \code{Y}, covariates \code{X},
logit regression coefficients \code{alpha} and a specified \code{model} of expert functions.
}
