% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict_excess_prior}
\alias{predict_excess_prior}
\title{predict_excess_prior}
\usage{
predict_excess_prior(X, alpha, model, limit, exposure_future = list())
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{model}{A matrix specifying the expert functions.}

\item{limit}{A matrix specifying the cutoff point.}

\item{exposure_future}{A vector indicating the time exposure (future) of each observation. If nothing is supplied, it is set to 1.0 by default.}
}
\value{
result: A matrix of predicted excess expectation of response, based on prior probabilities.
}
\description{
Predicts the excess expectation of response, given covariates \code{X},
logit regression coefficients \code{alpha} and a specified \code{model} of expert functions.
}
