% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict_mean_prior}
\alias{predict_mean_prior}
\title{predict_mean_prior}
\usage{
predict_mean_prior(X, alpha, model, exposure_future = list())
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{model}{An \link{ExpertMatrix}}

\item{exposure_future}{A vector indicating the time exposure (future) of each observation. If nothing is supplied,
it is set to 1.0 by default.}
}
\value{
result:
A matrix of predicted mean values of response, based on prior probabilities.
}
\description{
Predicts the mean values of response, given covariates \code{X}, logit regression coefficients \code{alpha} and a specified \code{model} of expert functions.
}
