% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict_var_prior}
\alias{predict_var_prior}
\title{predict_var_prior}
\usage{
predict_var_prior(X, alpha, model, exposure_future = list())
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{model}{A matrix specifying the expert functions.}

\item{exposure_future}{A vector indicating the time exposure of each observation. If nothing is supplied,
it is set to 1.0 by default.}
}
\value{
result A matrix of predicted variance of response, based on prior probabilities.
}
\description{
Predicts the variance of response, given covariates \code{X},
logit regression coefficients \code{alpha} and a specified \code{model} of expert functions.
}
