% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict_VaRCTE_posterior}
\alias{predict_VaRCTE_posterior}
\title{predict_VaRCTE_posterior}
\usage{
predict_VaRCTE_posterior(
  Y,
  X,
  alpha,
  model,
  p,
  exposure_past = list(),
  exposure_future = list(),
  exact_Y = F
)
}
\arguments{
\item{Y}{A matrix of responses.}

\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{model}{A matrix specifying the expert functions.}

\item{p}{A matrix of probabilities.}

\item{exposure_past}{A vector indicating the time exposure (past) of each observation. If nothing is supplied,it is set to 1.0 by default.}

\item{exposure_future}{A vector indicating the time exposure (future) of each observation. If nothing is supplied,it is set to 1.0 by default.}

\item{exact_Y}{Bool variable. indicating if \code{Y} is observed exactly or with censoring and truncation. Default set to be False}
}
\value{
result: list(\code{VaR}, \code{CTE})
\code{VaR}: A matrix of predicted VaR of response, based on posterior probabilities.
\code{CTE}: A matrix of predicted CTE of response, based on posterior probabilities.
}
\description{
Predicts the \code{p}-th value-at-risk (VaR) and conditional tail expectation (CTE) of response,
given observations \code{Y}, covariates \code{X}, logit regression coefficients \code{alpha} and a specified \code{model} of expert functions.
}
