% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict_VaRCTE_prior}
\alias{predict_VaRCTE_prior}
\title{predict_VaRCTE_prior}
\usage{
predict_VaRCTE_prior(X, alpha, model, p, exposure_future = list())
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{model}{A matrix specifying the expert functions.}

\item{p}{A matrix of probabilities.}

\item{exposure_future}{A vector indicating the time exposure (future) of each observation. If nothing is supplied,it is set to 1.0 by default.}
}
\value{
result: list(\code{VaR}, \code{CTE})
\code{VaR}: A matrix of predicted VaR of response, based on prior probabilities.
\code{CTE}: A matrix of predicted CTE of response, based on prior probabilities.
}
\description{
Predicts the \code{p}-th value-at-risk (VaR) and conditional tail expectation (CTE) of response,
given observations \code{Y}, covariates \code{X}, logit regression coefficients \code{alpha} and a specified \code{model} of expert functions.
}
