% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictClassPosterior.R
\name{PredictClassPosterior}
\alias{PredictClassPosterior}
\title{Predict the posterior latent class probabilities, given a fixed covariate matrix X, Y and a model.}
\usage{
PredictClassPosterior(
  Y,
  X,
  alpha,
  comp.dist,
  zero.prob,
  params.list,
  prob = TRUE
)
}
\arguments{
\item{Y}{A matrix of observed responses for \code{X}.}

\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}

\item{prob}{If \code{TRUE}, the function returns a matrix of latent class probabilities. If \code{FALSE},
it returns the index of the most likely latent class.}
}
\value{
A matrix of latent class probabilities by observation and by component.
}
\description{
Predict the posterior latent class probabilities, given a fixed covariate matrix X, Y and a model.
}
