% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictClassPrior.R
\name{PredictClassPrior}
\alias{PredictClassPrior}
\title{Predict the latent class probabilities, given a fixed covariate matrix X and a model.}
\usage{
PredictClassPrior(X, alpha, prob = TRUE)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{prob}{If \code{TRUE}, the function returns a matrix of latent class probabilities. If \code{FALSE},
it returns the index of the most likely latent class.}
}
\value{
A matrix of latent class probabilities by observation and by component.
}
\description{
Predict the latent class probabilities, given a fixed covariate matrix X and a model.
}
