% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictVariance.R
\name{PredictVariancePosterior}
\alias{PredictVariancePosterior}
\title{Predict the posterior variance of y, given a matrix of observations Y, a fixed covariate matrix X and a model.}
\usage{
PredictVariancePosterior(Y, X, alpha, comp.dist, zero.prob, params.list)
}
\arguments{
\item{Y}{A matrix of observed responses for \code{X}.}

\item{X}{A matrix of covariates.}

\item{alpha}{A matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, which specify the component distributions to fit.
The rows represent the dimensions of \code{Y}, while the columns represent the component distributions.}

\item{zero.prob}{A d*g matrix of numerics in (0,1), which specify the probability mass at zero for component distributions.}

\item{params.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the parameter value for the corresponding \code{comp.dist}.}
}
\value{
A matrix of variance by observation and by dimension.
}
\description{
Predict the posterior variance of y, given a matrix of observations Y, a fixed covariate matrix X and a model.
}
