% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{print_expert_matrix}
\alias{print_expert_matrix}
\title{Print expert matrix.}
\usage{
print_expert_matrix(expert_matrix)
}
\arguments{
\item{expert_matrix}{A matrix of expert functions}
}
\description{
Print expert matrix.
}
