% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{sim_exposurize_model}
\alias{sim_exposurize_model}
\title{Simulations of Exposurized Models}
\usage{
sim_exposurize_model(expert_matrix, exposure_list, selected_index)
}
\arguments{
\item{expert_matrix}{(\code{ExpertMatrix})\cr
The Expert Matrix that will be exposurized and simulated.}

\item{exposure_list}{(\code{numeric})\cr
The list of exposure values. Make sure this is a row vector with length to be \code{expert_matrix$ncol}}

\item{selected_index}{(\code{numeric})\cr
The list that indicate which latent class is simulated.}
}
\value{
result (\code{matrix})\cr
The result matrix with dimension to be \code{expert_matrix$nrow} rows and \code{expert_matrix$ncol} columns
}
\description{
Give one simulation for each expert in the expert_matrix after they got exposurized.
}
