% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchPosSim.R
\name{SimPosY}
\alias{SimPosY}
\title{A function to simulate positive part of observations y}
\usage{
SimPosY(sample.size.n, comp.dist, params)
}
\arguments{
\item{sample.size.n}{The size of y to simulate.}

\item{comp.dist}{A string which indicates the expert function.}

\item{params}{A vector of parameters for \code{ind.dist}.}
}
\description{
A function to simulate positive part of observations y
}
\keyword{internal}
