% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYInd.R
\name{SimYInd}
\alias{SimYInd}
\title{Simulate y, given a fixed covariate vector X and a model}
\usage{
SimYInd(X, alpha, comp.dist, zero.prob, params.list, sample.size = 1)
}
\arguments{
\item{X}{A P-element vector of covariates. The first element must be 1.}

\item{alpha}{A g*P matrix of logit regression coefficients.}

\item{comp.dist}{A d*g matrix of strings, describing component distributions by dimension and by component.}

\item{zero.prob}{A d*g matrix of numbers between 0 and 1, describing zero probability masses by dimension and by component.}

\item{sample.size}{Sample size to simulate. Default is 1.}

\item{paramas.list}{A list of length d, where each element is a sublist of length g.
Each sublist contains one numeric vector, which is the initial parameter guess for the corresponding \code{comp.dist}.}
}
\value{
A matrix of simulated values.
}
\description{
Simulate y, given a fixed covariate vector X and a model
}
\keyword{internal}
